+++ src/colorer/unicode/CString.h
@@ -40,7 +40,7 @@ public:
       @param def_encoding Default encoding to be used, if no other
              variants found.
   */
-  CString(const byte* stream, size_t size, int def_encoding = -1);
+  CString(const byte* stream, int32_t size, int def_encoding = -1);
 
   /** String from single-byte character buffer.
       @param string Character buffer, can't be null.
@@ -50,7 +50,7 @@ public:
       @param encoding Encoding, to use for char2unicode transformations.
              If -1, default encoding will be used.
   */
-  CString(const char* string, size_t s = 0, size_t l = npos, int encoding = -1);
+  CString(const char* string, int32_t s = 0, int32_t l = npos, int encoding = -1);
 
   /** String from unicode two-byte character buffer.
       @param string Unicode character buffer, can't be null.
@@ -58,17 +58,17 @@ public:
       @param l Length of created string. If npos, autodetects string length with
              last zero char.
   */
-  CString(const w2char* string, size_t s = 0, size_t l = npos);
+  CString(const w2char* string, int32_t s = 0, int32_t l = npos);
   /** String from UCS4 four-byte character buffer.
       @param string UCS4 unicode character buffer, can't be null.
       @param s Starting string position. Zero - create from start of buffer.
       @param l Length of created string. If npos, autodetects string length with
              last zero char.
   */
-  CString(const w4char* string, size_t s = 0, size_t l = npos);
+  CString(const w4char* string, int32_t s = 0, int32_t l = npos);
 
 #if (__WCHAR_MAX__ > 0xffff)
-  CString(const wchar* string, size_t s = 0, size_t l = npos);
+  CString(const wchar* string, int32_t s = 0, int32_t l = npos);
 #endif
 
   /** String from any @c String implementing interface.
@@ -77,7 +77,7 @@ public:
       @param l Length of created string. If npos, autodetects string length with
              cstring.length() call.
   */
-  CString(const String* cstring, size_t s = 0, size_t l = npos);
+  CString(const String* cstring, int32_t s = 0, int32_t l = npos);
 
   /** String from any @c String implementing interface.
       @param cstring String class instance.
@@ -85,7 +85,7 @@ public:
       @param l Length of created string. If npos, autodetects string length with
              cstring.length() call.
   */
-  CString(const String &cstring, size_t s = 0, size_t l = npos);
+  CString(const String &cstring, int32_t s = 0, int32_t l = npos);
 
   /** Empty String */
   CString();
@@ -95,8 +95,8 @@ public:
   CString(CString &&cstring);
   CString &operator=(CString &&cstring);
 
-  wchar operator[](size_t i) const override;
-  size_t length() const override;
+  wchar operator[](int32_t i) const override;
+  int32_t length() const override;
 
 protected:
   enum EStreamType {
@@ -118,8 +118,8 @@ protected:
     const String* cstr;
     wchar* stream_wstr;
   };
-  size_t start;
-  size_t len;
+  int32_t start;
+  int32_t len;
 
 };
